fc660c: Change default keymap and binary
