+++ keyboard/fc660c/binary/fc660c_emu.hex
@@ -148,9 +148,9 @@
 :10093000455920434F4E534F4C4520434F4D4D4159
 :100940004E44204E4B524F204B45594D41505F5322
 :10095000454354494F4E20343039360A00425549F8
-:100960004C443A20343466633536612D64697274C0
-:1009700079202830343A34303A35382041707220AA
-:1009800031312032303137290A005649443A20307B
+:100960004C443A20653338323935372D6469727416
+:1009700079202831363A31333A32362041707220AC
+:1009800031352032303137290A005649443A203077
 :10099000784645454428544D4B2E29205049443A29
 :1009A00020307836363043284643363630432041AF
 :1009B0006C7420436F6E74726F6C6C657229205674
