--- keyboard/fc660c/unimap_plain.c
@@ -1,19 +1,0 @@
-#include "action.h"
-#include "actionmap.h"
-#include "hook.h"
-#include "fc660c.h"
-
-
-#ifdef KEYMAP_SECTION_ENABLE
-const action_t actionmaps[][UNIMAP_ROWS][UNIMAP_COLS] __attribute__ ((section (".keymap.keymaps"))) = {
-#else
-const action_t actionmaps[][UNIMAP_ROWS][UNIMAP_COLS] PROGMEM = {
-#endif
-    [0] = KMAP(
-        ESC, 1,   2,   3,   4,   5,   6,   7,   8,   9,   0,   MINS,EQL, BSPC,     INS,
-        TAB, Q,   W,   E,   R,   T,   Y,   U,   I,   O,   P,   LBRC,RBRC,BSLS,     DEL,
-        CAPS,A,   S,   D,   F,   G,   H,   J,   K,   L,   SCLN,QUOT,     ENT,
-        LSFT,Z,   X,   C,   V,   B,   N,   M,   COMM,DOT, SLSH,          RSFT,UP,
-        LCTL,LGUI,LALT,          SPC,                     RALT,RCTL,RGUI,LEFT,DOWN,RGHT
-    ),
-};
